+++ packages/pluggableWidgets/combobox-web/src/__tests__/MultiSelection.spec.tsx
@@ -9,7 +9,7 @@ import {
 import "./__mocks__/intersectionObserverMock";
 import "@testing-library/jest-dom";
 import { fireEvent, render, RenderResult, waitFor } from "@testing-library/react";
-import { ObjectItem, DynamicValue } from "mendix";
+import { ObjectItem, DynamicValue, ListValue } from "mendix";
 import { createElement } from "react";
 import { ComboboxContainerProps } from "../../typings/ComboboxProps";
 import Combobox from "../Combobox";
@@ -47,6 +47,7 @@ describe("Combo box (Association)", () => {
             selectedItemsStyle: "text",
             readOnlyStyle: "bordered",
             lazyLoading: false,
+            loadingType: "spinner",
             noOptionsText: dynamicValue("no options found"),
             clearButtonAriaLabel: dynamicValue("Clear selection"),
             removeValueAriaLabel: dynamicValue("Remove value"),
@@ -158,5 +159,33 @@ describe("Combo box (Association)", () => {
         expect(defaultProps.attributeAssociation?.value).toHaveLength(1);
         fireEvent.click(selectAllButton);
         expect(defaultProps.attributeAssociation?.value).toHaveLength(4);
+    });
+
+    describe("with lazy loading", () => {
+        it("calls loadMore only when menu opens", async () => {
+            const setLimit = jest.fn();
+            const lazyLoadingProps = {
+                ...defaultProps,
+                lazyLoading: true,
+                optionsSourceAssociationDataSource: {
+                    ...defaultProps.optionsSourceAssociationDataSource,
+                    hasMoreItems: true,
+                    limit: 0,
+                    setLimit
+                } as ListValue
+            };
+            const component = render(<Combobox {...lazyLoadingProps} />);
+
+            expect(component.queryAllByRole("option")).toHaveLength(0);
+            expect(lazyLoadingProps.optionsSourceAssociationDataSource?.limit).toEqual(0);
+
+            const input = await getInput(component);
+            fireEvent.click(input);
+
+            await waitFor(() => {
+                expect(component.queryAllByRole("option")).toHaveLength(4);
+                expect(setLimit).toHaveBeenCalledWith(100);
+            });
+        });
     });
 });