+++ packages/pluggableWidgets/combobox-web/src/__tests__/StaticSelection.spec.tsx
@@ -46,6 +46,7 @@ describe("Combo box (Static values)", () => {
             selectedItemsStyle: "text",
             readOnlyStyle: "bordered",
             lazyLoading: false,
+            loadingType: "spinner",
             clearButtonAriaLabel: dynamicValue("Clear selection"),
             removeValueAriaLabel: dynamicValue("Remove value"),
             selectAllButtonCaption: dynamicValue("Select All"),